Workflow/backend/signature/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
tensorflow
ultralytics
huggingface_hub
pillow
numpy
